Fixed some errors / added some features: - The old GUI doesn't print out no more on the left hand side - You can now, if your party isn't full, click on an animal by the bush and add them into the party instead of animal swapping - Getting caught now renders the mouse cursor to the flamingo state

+++ trunk/prog/src/org/cart/igd/game/Inventory.java
@@ -41,7 +41,7 @@ public class Inventory{
 	  ///
 	  public int PSYCH_FOOD_WATER_AFFINITY = 0; // 1 is food 2 is water
 	  ///
-	  public long PSYCH_TIME_IN_UNIMPORTANT_PLACES_ON_MAP = 0;
+	  public int PSYCH_TIME_IN_UNIMPORTANT_PLACES_ON_MAP = 0;
 	  public int PSYCH_ENTERED_UNIMPORTANT_PLACES_ON_MAP = 0;
 	  ///
 	  public int PSYCH_UNNECESSARY_CLICKS = 0;
